Ethernet is considered the first header to be parsed in Pcapplusplus and Scapy
There is a variable called "START_WITH_ETH" in p4-traffictool.sh. By setting this variable as true, source codes for headers before or at the same level of Ethernet won't be generated.
The changes are in DissectTrafficPcap.py and GenTrafficScapy.py. A Class called 'State' is added to represent each header type in the graph. Two methods, 'make_tree' and 'find_children'
are used to create the headers' trees. 'find_ethernet' and 'find_eth_subhdr' are used to specifically define wich headers are reached through ethernet header path.